annoy-be/requirements.txt
annoy==1.17.3
Flask
google-analytics-data==0.17.1